(express) UPDATE(put, patch), DELETE EXAMPLE
