+++ dataMocking/src/cats/cats.model.ts
